Version 0.0.3 from OGC's gitlab repo
This is the first complete draft from Fuhu Ren's team. The draft was provide as Word docx file, which has been converted to metanorma asciidoc, and compiled to this file. Issues raised from now against this file, will be managed in github with associated changes tracked in the Editor's OGC gitlab source. Issues will be closed when a new pdf file is carried over from the gitlab.